# backend/src/products/products.controller.ts
import { Controller, Get, Post, Put, Delete, Param, Query, Body } from '@nestjs/common';
import { ApiTags, ApiOperation, ApiQuery } from '@nestjs/swagger';
import { ProductsService } from './products.service';
import { CreateProductDto, UpdateProductDto } from './dto/product.dto';

@ApiTags('products')
@Controller('products')
export class ProductsController {
  constructor(private readonly productsService: ProductsService) {}

  @Get()
  @ApiOperation({ summary: 'Get products with pagination and filters' })
  @ApiQuery({ name: 'shopId', required: true })
  @ApiQuery({ name: 'categoryIds', required: false, description: 'Comma-separated category IDs' })
  @ApiQuery({ name: 'page', required: false, type: Number })
  @ApiQuery({ name: 'limit', required: false, type: Number })
  @ApiQuery({ name: 'sortBy', required: false, enum: ['name', 'price', 'createdAt'] })
  @ApiQuery({ name: 'sortDir', required: false, enum: ['asc', 'desc'] })
  findAll(
    @Query('shopId') shopId: string,
    @Query('categoryIds') categoryIds?: string,
    @Query('page') page = 1,
    @Query('limit') limit = 6,
    @Query('sortBy') sortBy = 'name',
    @Query('sortDir') sortDir: 'asc' | 'desc' = 'asc',
  ) {
    const ids = categoryIds ? categoryIds.split(',').filter(Boolean) : undefined;
    return this.productsService.findAll({
      shopId,
      categoryIds: ids,
      page: +page,
      limit: +limit,
      sortBy,
      sortDir,
    });
  }

  @Get(':id')
  @ApiOperation({ summary: 'Get product by ID' })
  findOne(@Param('id') id: string) {
    return this.productsService.findOne(id);
  }

  @Post()
  @ApiOperation({ summary: 'Create a new product' })
  create(@Body() dto: CreateProductDto) {
    return this.productsService.create(dto);
  }

  @Put(':id')
  @ApiOperation({ summary: 'Update a product' })
  update(@Param('id') id: string, @Body() dto: UpdateProductDto) {
    return this.productsService.update(id, dto);
  }

  @Delete(':id')
  @ApiOperation({ summary: 'Delete a product' })
  remove(@Param('id') id: string) {
    return this.productsService.remove(id);
  }
}