# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/shops": {
      "get": {
        "operationId": "ShopsController_findAll",
        "summary": "Get all shops with optional rating filter",
        "parameters": [
          {
            "name": "minRating",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxRating",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "shops"
        ]
      },
      "post": {
        "operationId": "ShopsController_create",
        "summary": "Create a new shop",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShopDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "shops"
        ]
      }
    },
    "/shops/{id}": {
      "get": {
        "operationId": "ShopsController_findOne",
        "summary": "Get shop by ID with products",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "shops"
        ]
      },
      "put": {
        "operationId": "ShopsController_update",
        "summary": "Update a shop",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateShopDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "shops"
        ]
      },
      "delete": {
        "operationId": "ShopsController_remove",
        "summary": "Delete a shop",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "shops"
        ]
      }
    },
    "/products": {
      "get": {
        "operationId": "ProductsController_findAll",
        "summary": "Get products with pagination and filters",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryIds",
            "required": false,
            "in": "query",
            "description": "Comma-separated category IDs",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDir",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "name",
                "price",
                "createdAt"
              ],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products"
        ]
      },
      "post": {
        "operationId": "ProductsController_create",
        "summary": "Create a new product",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "summary": "Get product by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products"
        ]
      },
      "put": {
        "operationId": "ProductsController_update",
        "summary": "Update a product",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "summary": "Delete a product",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "products"
        ]
      }
    },
    "/orders": {
      "post": {
        "operationId": "OrdersController_create",
        "summary": "Create a new order with items",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/search": {
      "get": {
        "operationId": "OrdersController_search",
        "summary": "Search orders by email and phone",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "summary": "Get order by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "orders"
        ]
      }
    },
    "/coupons": {
      "get": {
        "operationId": "CouponsController_findAllActive",
        "summary": "Get all active coupons",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "coupons"
        ]
      },
      "post": {
        "operationId": "CouponsController_create",
        "summary": "Create a new coupon",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCouponDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "coupons"
        ]
      }
    },
    "/coupons/all": {
      "get": {
        "operationId": "CouponsController_findAll",
        "summary": "Get all coupons (including inactive)",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "coupons"
        ]
      }
    },
    "/coupons/validate/{code}": {
      "get": {
        "operationId": "CouponsController_validate",
        "summary": "Validate a coupon code",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "coupons"
        ]
      }
    },
    "/coupons/{id}": {
      "get": {
        "operationId": "CouponsController_findOne",
        "summary": "Get coupon by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "coupons"
        ]
      },
      "put": {
        "operationId": "CouponsController_update",
        "summary": "Update a coupon",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCouponDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "coupons"
        ]
      },
      "delete": {
        "operationId": "CouponsController_remove",
        "summary": "Delete a coupon",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "coupons"
        ]
      }
    },
    "/categories": {
      "get": {
        "operationId": "CategoriesController_findAll",
        "summary": "Get all categories",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "categories"
        ]
      },
      "post": {
        "operationId": "CategoriesController_create",
        "summary": "Create a new category",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "categories"
        ]
      }
    },
    "/categories/{id}": {
      "get": {
        "operationId": "CategoriesController_findOne",
        "summary": "Get category by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "categories"
        ]
      },
      "put": {
        "operationId": "CategoriesController_update",
        "summary": "Update a category",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "categories"
        ]
      },
      "delete": {
        "operationId": "CategoriesController_remove",
        "summary": "Delete a category",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "categories"
        ]
      }
    }
  },
  "info": {
    "title": "Meal Magic Delivery API",
    "description": "REST API for shops, products, coupons and orders",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "shops",
      "description": ""
    },
    {
      "name": "products",
      "description": ""
    },
    {
      "name": "categories",
      "description": ""
    },
    {
      "name": "orders",
      "description": ""
    },
    {
      "name": "coupons",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "CreateShopDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Best Burgers"
          },
          "imageUrl": {
            "type": "string"
          },
          "rating": {
            "type": "number",
            "example": 4.5
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateShopDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "rating": {
            "type": "number"
          }
        }
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Big Burger"
          },
          "price": {
            "type": "number",
            "example": 9.99
          },
          "imageUrl": {
            "type": "string"
          },
          "shopId": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "discountPercent": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "name",
          "price",
          "shopId",
          "categoryId"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "imageUrl": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "discountPercent": {
            "type": "number"
          }
        }
      },
      "OrderItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "price": {
            "type": "number"
          }
        },
        "required": [
          "productId",
          "quantity",
          "price"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "totalPrice": {
            "type": "number"
          },
          "couponCode": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDto"
            }
          }
        },
        "required": [
          "email",
          "phone",
          "address",
          "totalPrice",
          "items"
        ]
      },
      "CreateCouponDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Summer Sale"
          },
          "code": {
            "type": "string",
            "example": "SUMMER20"
          },
          "discountPercent": {
            "type": "number",
            "example": 20
          },
          "imageUrl": {
            "type": "string"
          },
          "active": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "name",
          "code",
          "discountPercent"
        ]
      },
      "UpdateCouponDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "discountPercent": {
            "type": "number"
          },
          "imageUrl": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Burgers"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Pizza"
          }
        }
      }
    }
  }
}